Make lemem work with cgroups v2
This does the necessary to make lemem work with cgroups v2, if you want
cgroups v1 then checkout the 'cgroup_v1' tag.

The big difference with v2 is that all the controllers are just under a
unified hierarchy, rather than having separate directories for
cpu,memory,etc it's just under a single /sys/fs/cgroup directory
structure.

The control file names have changed. It also seems we don't need to do
the cleanup as we did previously. Also how the swap usage is controlled
has changed, it's now treated as a separate resource and if you use the
-s option, it simply disables the use of swap for the process.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>